+++ SOTFEdit/Model/ScrewStructureModificationMode.cs
@@ -4,6 +4,7 @@ public enum ScrewStructureModificationMode
 {
     None,
     Remove,
+    AlmostFinish,
     Finish,
     Unfinish
 }