+++ SOTFEdit/Model/ScrewStructureModificationWrapper.cs
@@ -1,0 +1,7 @@
+﻿namespace SOTFEdit.Model;
+
+public class ScrewStructureModificationWrapper
+{
+    public ScrewStructureModificationMode? ModificationMode { get; set; }
+    public ScrewStructureOrigin Origin { get; set; }
+}