+++ SOTFEdit/Model/ScrewStructureOrigin.cs
@@ -1,0 +1,7 @@
+﻿namespace SOTFEdit.Model;
+
+public enum ScrewStructureOrigin
+{
+    Unfinished,
+    Finished
+}