+++ SOTFEdit/Model/UnfinishedGenericScrewStructure.cs
@@ -1,0 +1,5 @@
+﻿using Newtonsoft.Json.Linq;
+
+namespace SOTFEdit.Model;
+
+public record UnfinishedGenericScrewStructure(int Id, Position Pos, JToken Rot, int Added);